+++ src/error/mod.rs
@@ -9,6 +9,7 @@ use std::path;
 mod bad_input;
 mod cell_parse_error;
 mod display;
+mod eval_error;
 mod from;
 mod parse_error;
 
@@ -16,6 +17,9 @@ pub type Result<T> = std::result::Result<T, Error>;
 
 pub(crate) use bad_input::BadInput;
 pub(crate) use cell_parse_error::CellParseError;
+
+pub use eval_error::EvalError;
+pub(crate) type EvalResult<T> = std::result::Result<T, EvalError>;
 
 pub use parse_error::ParseError;
 pub(crate) type ParseResult<T> = std::result::Result<T, ParseError>;
@@ -39,9 +43,9 @@ pub enum Error {
 
     /// An error encountered when evaluating the formulas in a cell.
     EvalError {
-        message: String,
+        eval_error: Box<EvalError>,
         filename: path::PathBuf,
-        position: a1_notation::Address,
+        position: Option<a1_notation::Address>,
     },
 
     /// Google Sheets requires that the `gcloud` CLI tools are installed and configured.  If we