+++ src/module_loader/dependency.rs
@@ -1,4 +1,4 @@
-use crate::Scope;
+use crate::{ArcSourceCode, Scope};
 
 #[derive(Copy, Clone, Debug)]
 pub(super) enum DependencyRelation {
@@ -10,22 +10,5 @@ pub(super) enum DependencyRelation {
 pub(super) struct Dependency {
     pub(super) relation: DependencyRelation,
     pub(super) scope: Scope,
-}
-
-impl Dependency {
-    #[cfg(test)]
-    pub(super) fn direct(scope: Scope) -> Self {
-        Self {
-            relation: DependencyRelation::Direct,
-            scope,
-        }
-    }
-
-    #[cfg(test)]
-    pub(super) fn transitive(scope: Scope) -> Self {
-        Self {
-            relation: DependencyRelation::Transitive,
-            scope,
-        }
-    }
+    pub(super) source_code: ArcSourceCode,
 }