+++ docs/CHANGELOG.md
@@ -7,6 +7,10 @@
 * Code coverage reporting
 * Tooling to run benchmarks
 
+### Bugfixes
+
+* Fix unfriendly error when calling a function with the wrong number of arguments
+
 ### **Breaking (Language) Changes**
 
 * All builtin variables & functions are removed.  All of the functionality provided by them could be done with native spreadsheet functions anyway.  And we're adding support for module loading which means they could just be implemented as a lib instead.
@@ -16,6 +20,7 @@
 * Rename `CodeSection` to `Scope`
 * Rename `Error::ObjectWriteError` to `Error::ObjectCodeError`
 * `Template.write_object_file` made crate-private
+* `Error::EvalError` changed
 
 ## v0.4.1